+++ src/test/formatter_pint_php_blocks.test.ts
@@ -497,5 +497,40 @@ fn () => true;
 </div>
 `
         assert.strictEqual(formatBladeStringWithPint(input), out);
-    })
+    });
+
+    test('pint: it automatically removes strict types', () => {
+        const input =  `
+
+@if (! $something)
+
+@endif
+
+@php
+    $size = match ($size) {
+        'xs' => 'h-5 w-5 md:h-4 md:w-4',
+        'sm' => 'h-6 w-6 md:h-5 md:w-5',
+        'md' => 'h-7 w-7 md:h-6 md:w-6',
+        'lg' => 'h-8 w-8 md:h-7 md:w-7',
+        'xl' => 'h-9 w-9 md:h-8 md:w-8',
+        default => $size,
+    };
+@endphp
+`;
+        const expected = `@if (! $something)
+@endif
+
+@php
+    $size = match ($size) {
+        'xs' => 'h-5 w-5 md:h-4 md:w-4',
+        'sm' => 'h-6 w-6 md:h-5 md:w-5',
+        'md' => 'h-7 w-7 md:h-6 md:w-6',
+        'lg' => 'h-8 w-8 md:h-7 md:w-7',
+        'xl' => 'h-9 w-9 md:h-8 md:w-8',
+        default => $size,
+    };
+@endphp
+`;
+        assert.strictEqual(formatBladeStringWithPint(input), expected);
+    });
 });